+++ Net23Online/WebNet23Online/Models/DelightBistro/CreateIngredientViewModel.cs
@@ -9,9 +9,9 @@ namespace WebNet23Online.Models.DelightBistro
         [Required]
         //[IsUniqueIngredient] // ошибка при передаче Name при создании/обновлении блюда
         public string Name { get; set; } 
-        public int Quantity { get; set; } = 10;
+        public decimal Quantity { get; set; } = 10;
         [Range(1, 1000)]
-        public int Price { get; set; } = 1;
+        public decimal Price { get; set; } = 1;
         public bool IsSelected { get; set; } = false;
     }
 }
